// frontend/src/lib/types/Message.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SystemEvent } from "./SystemEvent";

/**
 * タグ付きメッセージ種別
 */
export type Message = { "type": "chat", content: string, } | { "type": "system", event: SystemEvent, message: string, };